update look and fix deleting projects

+++ Angular/TunRW-app/backend/controller/projects.js
@@ -135,7 +135,7 @@ exports.createProject = (req,res,next) => {
         watched : documents,
         creator : req.userData.username
       });
-      Project.deleteOne().then(result =>{
+      Project.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result =>{
         if(result.n > 0){
           notification.save().then(notResult => {
             res.status(200).json({
